SR-dev/requirements.txt
rdflib>=6.0.0
requests>=2.31.0
SPARQLWrapper>=2.0.0
PyYAML>=6.0.1 